requirements.txt
Flask==2.2.3
pymongo[srv]==4.5.0
python-dotenv==1.0.0
